requirements.txt
pyvis
dss-python[all]
pywin32